inline: idea for a nextstate after the inlined call
when no enter/leave was added,
to fix the oplines for the next expression.
otherwise we continue with the wrong line from the inlines CV
until the next nextstate.
But we cannot do this, since nextstate resets the SP, so
the next op will have a wrong stack, so we would need ENTER/LEAVE
so save/restore this properly. So go with wrong lines.