native: upgrade padsv at run-time
Upgrade a OP_PADSV op to its native variant, and mark it as BOXRET,
pushing a boxed SV onto the stack.
This is needed because different padsv ops can point to the same
native curpad entry, which could have been upgraded, but the ops not.

Do this on the fly in pp_padsv, so that the next call will use the faster
variant. This indirectly promotes the lexical variable to a native type,
but uses the non-native context, because of the BOXRET flag.

Provide SvNATIVE backcompat for non-usecperl.