symtab: gv_search_symtab flat table
With search, insert and delete actions for GVs.
We just rely on the hash, and store the name in the value.
See [cperl #127]

todo: utf8 in name, name check when found,
grow, wrap search around, cuckoo (for now just linear).
integrate.

problematic to refactor: defstash, curstash, (debstash).
they can point to symtab entries containing the stash name,
like "main::" => NAME (gv or hv), just for the current package
name. but changing those stashes and searching is useless, so the value
doesn't need to be a full HV or GV.