lazyparse: #274 WIP
on SUB subname scan the block ahead and store it
as string SV in CvLAZY(PL_compcv).
Don't store any optree, defer it to later when the sub
is actually called.
This is a benefit when there's a lot of dead code (~70% typically)
and the storage to store the string buffers is not too big.
i.e. many ops or 1-3 ops.

Better than storing the block buffer would be just to store just
the mmap'ed bufptr, offset and len of the source file. But we don't
mmap by default yet, even if it's faster.

Note that subs are stored in slabs. As we store almost no slabs at
compile-time (via PL_compcv) it gets tricky.