preserve numericness of system() args on Win32
On Windows there's a nasty variation in the meaning of arguments
to Perl's system(), in which a numeric first argument isn't used as
part of the command to run, but instead selects between two different
operations to perform with the command (whether to wait for the command
to complete or not).  Therefore the reduction of argument scalars to
their operative values in the parent process, which was added in commit
64def2aeaeb63f92dadc6dfa33486c1d7b311963, needs to preserve numericness
of arguments on Windows.  Fixes [perl #132633].

(cherry picked from commit 8fe3452cc6ac7af8c08c2044cd3757018a9c8887)

[cperl: this introduces calling get-magic for all system args twice
 and does not fix VMS breakage.]