vivify array elements when putting them on stack
When the elements of an array are put on the stack, by a padav, rv2av,
or padrange op, null pointers in the AvARRAY were being pushed as
&PL_sv_undef, which was OK in rvalue contexts but caused misbehaviour
in lvalue contexts.  Change this to vivify these elements.  There's no
attempt here to limit the vivification to lvalue contexts: the existing
op flags aren't enough to detect \(@a), and attempting to catch all
cases where a new flag needs to be set would be an error-prone process.
Fixes [perl #8910].

(cherry picked from commit 6661956a23de82b41adc406200054293d6d7aded)

[cperl: I'm not convinced. This seems to be a huge perf. regression,
and catching all lval contexts seems to be better, than creating a penalty
for everyone. This is used for padrange: my (.., ..),
padav (lex. arrays) and rv2av. But actually I cannot measure a critical
slowdown, so I let it in. Maybe my benchmarks don't cover it.]