grok_numeric_radix(): Avoid recalculating
This function just determined that we are in the scope of 'use locale',
hence the underlying radix character should be used.  This commit
changes to use the macro that directly does that; previously the macro
that redundantly looks at if we are in the scope was used.

(cherry picked from commit a1395eaf50e445a5d48b26b960d58275dcffc265)