isSCRIPT_RUN: Can short cut if not in UTF-8
All characters representable by single bytes are either Common or Latin,
so must be a script run.  If we aren't asking for what the script is we
can return immediately.  If we are, the run is Latin if any character in
it is Latin, otherwise is Common.

(cherry picked from commit eac3f4d5c7d71a1a3a22818c6bcc9c32b312eb40)