Follow-up to fd77b29b3be4
As Zefram pointed out, I left in a piece of code that caused one
branch to continue to behave as before.  The change was ineffective
and the tests happened to be written in such a way as to take the
other branch.

(cherry picked from commit 0f3591f94499f7771d7a8e07afb6776a93fc5f9f)