regexec.c: Don't retest the same byte immediately
In this macro, COND has just returned true for the given byte.  We then
need to test that the rest of the relevant portion of the input string
and pattern match.  But before this commit, we started at the byte we
already know the answer for.  Change to test starting one position over.

(cherry picked from commit 8b37b57be5561e87de1003b536c9dfbfd6902d10)